+++ tests/kit/model_zoo/torchrec/__init__.py
@@ -1,1 +1,1 @@
-from .torchrec import *
+#from .torchrec import *
